Add e2e test for GenericRefField recent fix
